food allergy test
Food is certainly no direct threat to the organism. In contrast, it is necessary for a
sufficient supply of energy, nutrients and micronutrients, to allow normal functions
of a living cell or the organism. However, when symptom’s like headache, migraine,
bloating, acidity build up it is necessity to do a Food Allergy Test.
Enteric cells are glued together by the so-called tight junctions. They prevent food
from by-passing the cells and force them to enter the organism by passing through
the cells. If the gut barrier is impaired, the tight junctions may break up, increasing
the permeability of the gut and leading to leaky gut syndrome. A food Allergy Test is
done to identify which food is responsible for the development.